Changed major computations in the controller to doubles, added a low pass filter on wd.  I just had the first working version that was very smooth.  I have yet to evaluate its effectiveness.  I need to bag some files and make plots, or possibly just use the camera system.  Note that this version could use a bit of cleanup as well.  Left come comments in so that I have them as a reference in this commit.

+++ prototypes_al_pupp.h
@@ -51,7 +51,7 @@ void CreateAndSendArray(unsigned short id, unsigned char *DataString);
 void delay(void);
 
 // Functions for kinematic controller:
-void run_filo(const float new_val, float *array);
+void run_filo(const double new_val, double *array);
 void calculate_controller_gains(void);
 int calculate_feedforward_values(const float k);
 void setup_controller(void);
